Python/AI-CHATTER_task4/requirements.txt
Flask
Flask_SQLAlchemy
python-dotenv
requests
gunicorn
